[Frontend][Rust] Move model-specific reasoning tests into their modules
