+++ rust/src/parser/src/reasoning/deepseek_r1.rs
@@ -41,3 +41,21 @@ impl ReasoningParser for DeepSeekR1ReasoningParser {
         Ok(self.inner.finish())
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use super::DeepSeekR1ReasoningParser;
+    use crate::reasoning::tests::{content_str, fake_tokenizer, push_str};
+
+    #[test]
+    fn deepseek_r1_without_prompt_markers_expects_start_token() {
+        let tokenizer = Arc::new(fake_tokenizer());
+        let mut parser = DeepSeekR1ReasoningParser::new(tokenizer).unwrap();
+
+        let delta = push_str(&mut parser, "reason</think>answer");
+        assert_eq!(delta.reasoning, None);
+        assert_eq!(content_str(&delta), Some("reason</think>answer"));
+    }
+}
