+++ rust/src/parser/src/reasoning/qwen3.rs
@@ -44,3 +44,64 @@ impl ReasoningParser for Qwen3ReasoningParser {
         Ok(self.inner.finish())
     }
 }
+
+#[cfg(test)]
+mod tests {
+    use std::sync::Arc;
+
+    use super::Qwen3ReasoningParser;
+    use crate::reasoning::ReasoningParser;
+    use crate::reasoning::tests::{
+        SPECIAL_BOUNDARY_ID, THINK_END_ID, THINK_START_ID, content_str, fake_tokenizer, push_str,
+        reasoning_str,
+    };
+
+    #[test]
+    fn qwen3_without_prompt_markers_expects_start_token() {
+        let tokenizer = Arc::new(fake_tokenizer());
+        let mut parser = Qwen3ReasoningParser::new(tokenizer).unwrap();
+
+        let delta = push_str(&mut parser, "reason</think>answer");
+        assert_eq!(delta.reasoning, None);
+        assert_eq!(content_str(&delta), Some("reason</think>answer"));
+    }
+
+    #[test]
+    fn qwen3_prompt_end_marker_starts_in_content() {
+        let tokenizer = Arc::new(fake_tokenizer());
+        let mut parser = Qwen3ReasoningParser::new(tokenizer).unwrap();
+        parser.initialize(&[THINK_END_ID]).unwrap();
+
+        let delta = push_str(&mut parser, "answer");
+        assert_eq!(delta.reasoning, None);
+        assert_eq!(content_str(&delta), Some("answer"));
+    }
+
+    #[test]
+    fn qwen3_tolerates_old_and_new_formats() {
+        let tokenizer = Arc::new(fake_tokenizer());
+
+        let mut old_parser = Qwen3ReasoningParser::new(tokenizer.clone()).unwrap();
+        let old = push_str(&mut old_parser, "<think>reason</think>answer");
+        assert_eq!(reasoning_str(&old), Some("reason"));
+        assert_eq!(content_str(&old), Some("answer"));
+
+        let mut new_parser = Qwen3ReasoningParser::new(tokenizer).unwrap();
+        new_parser.initialize(&[THINK_START_ID]).unwrap();
+        let new = push_str(&mut new_parser, "reason</think>answer");
+        assert_eq!(reasoning_str(&new), Some("reason"));
+        assert_eq!(content_str(&new), Some("answer"));
+    }
+
+    #[test]
+    fn qwen3_stops_scanning_at_last_special_token() {
+        let tokenizer = Arc::new(fake_tokenizer());
+        let mut parser = Qwen3ReasoningParser::new(tokenizer).unwrap();
+
+        parser.initialize(&[THINK_START_ID, SPECIAL_BOUNDARY_ID]).unwrap();
+
+        let delta = push_str(&mut parser, "answer");
+        assert_eq!(delta.reasoning, None);
+        assert_eq!(content_str(&delta), Some("answer"));
+    }
+}
